pyproject.toml
[tool.poetry]
name = "artham-server"
version = "0.1.0"
description = ""
authors = ["Vinoth Pandian <vinothpandian@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0"
djangorestframework = "^3.13.1"
django-environ = "^0.8.1"
psycopg2-binary = "^2.9.2"

[tool.poetry.dev-dependencies]
rope = "^0.22.0"
flake8 = "^4.0.1"
black = "^21.12b0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length=79